replace get id with reference_id

+++ app/Http/Controllers/Api/V1/GroupController.php
@@ -49,8 +49,14 @@ class GroupController extends ApiController
     /**
      * Display the specified resource.
      */
-    public function show(Group $group)
+    public function show($reference_id)
     {
+        $group = Group::where('reference_id', $reference_id)->first();
+
+        if(!$group){
+            return $this->error('Group not found', 404);
+        }
+
         if(Gate::authorize('show-group', $group)) {
 
             if($this->include('members')) {
@@ -68,8 +74,14 @@ class GroupController extends ApiController
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateGroupRequest $request, Group $group)
+    public function update(UpdateGroupRequest $request, $reference_id)
     {
+        $group = Group::where('reference_id', $reference_id)->first();
+
+        if(!$group){
+            return $this->error('Group not found', 404);
+        }
+
         if(Gate::authorize('update-group', $group)) {
 
             $group->update($request->mappedAttributes());
@@ -81,8 +93,15 @@ class GroupController extends ApiController
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Group $group)
+    public function destroy($reference_id)
     {
+
+        $group = Group::where('reference_id', $reference_id)->first();
+
+        if(!$group){
+            return $this->error('Group not found', 404);
+        }
+
         if(Gate::authorize('delete-group', $group)) {
 
             foreach($group->payments as $payment){
@@ -98,23 +117,31 @@ class GroupController extends ApiController
         }
     }
 
-    public function add_members(Request $request, Group $group)
+    public function add_members(Request $request, $reference_id)
     {
+        $group = $this->get_group($reference_id);
+
         if(Gate::authorize('member-group', $group)){
             $member_ids = $request->input('data.attributes.member_ids');
             $group->members()->syncWithoutDetaching($member_ids);
 
-            return $this->ok('Members added');
+            return $this->ok('Members added', $member_ids);
         }
     }
 
-    public function remove_members(Request $request, Group $group)
+    public function remove_members(Request $request, $reference_id)
     {
+        $group = Group::where('reference_id', $reference_id)->first();
+
+        if(!$group){
+            return $this->error('Group not found', 404);
+        }
+
         if(Gate::authorize('member-group', $group)){
             $member_ids = $request->input('data.attributes.member_ids');
             $group->members()->detach($member_ids);
 
-            return $this->ok('Members removed');
+            return $this->ok('Members removed', $member_ids);
         }
     }
